requirements.txt
numpy==1.17.4
opencv-python==4.1.2.30
Pillow==6.2.1
torch==1.0.0
torch-encoding==1.0.1
torchvision==0.2.1
tqdm==4.40.2
